requirements.txt
django==4.2.7
djangorestframework==3.14.0
dj-database-url==1.3.0
gunicorn==20.1.0
whitenoise==6.4.0
django-cors-headers==3.14.0
dj-rest-auth[with_social]==3.0.0
djangorestframework-simplejwt==5.2.2
